KNOX-3103 - Move RegExUtils testing to an appropriate suite in the the utils module
